backends: pass only initialized data to nsswitch gethostbyname4_r
1) Initualize the result pointer to NULL before passing a pointer to it

This avoids memory corruption caused by _nss_files_gethostbyname4_r
which for some unknown reason attempts to use the random memory
referenced from the pointer that is supposed to be the output
parameter.

2) Initialize the TTL value before passing a pointer to it

I'm not yet sure what is the correct behavior regarding TTL in the
nsswitch interface. It seems that nscd/aicache.c in glibc source
code uses INT32_MAX and expects the module to clamp it down if
possible. But that belongs to a different commit.